Put the entities in a separate package (JPA).
