+++ freedom-of-travel-api/src/main/java/nc/project/const_enum/ERole.java
@@ -1,4 +1,4 @@
-package nc.project.models;
+package nc.project.const_enum;
 
 public enum ERole {
     ROLE_USER,
