+++ freedom-of-travel-api/src/main/java/nc/project/jpa/entity/UserRole.java
@@ -1,4 +1,4 @@
-package nc.project.models;
+package nc.project.jpa.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
