+++ freedom-of-travel-api/src/main/java/nc/project/jpa/repository/RoleRepository.java
@@ -1,7 +1,7 @@
-package nc.project.repository;
+package nc.project.jpa.repository;
 
-import nc.project.models.Role;
-import nc.project.models.ERole;
+import nc.project.jpa.entity.Role;
+import nc.project.const_enum.ERole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
