+++ freedom-of-travel-api/src/main/java/nc/project/jpa/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package nc.project.repository;
+package nc.project.jpa.repository;
 
-import nc.project.models.User;
+import nc.project.jpa.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
