+++ freedom-of-travel-api/src/main/java/nc/project/jpa/repository/UserRoleRepository.java
@@ -1,6 +1,6 @@
-package nc.project.repository;
+package nc.project.jpa.repository;
 
-import nc.project.models.UserRole;
+import nc.project.jpa.entity.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
