+++ freedom-of-travel-impl/src/main/java/nc/project/implement/TestDataLoaderImpl.java
@@ -3,9 +3,10 @@ package nc.project.implement;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nc.project.const_enum.ApartmentType;
+import nc.project.const_enum.ERole;
 import nc.project.const_enum.ReservationStatus;
-import nc.project.models.*;
-import nc.project.repository.*;
+import nc.project.jpa.entity.*;
+import nc.project.jpa.repository.*;
 import nc.project.service.TestDataLoader;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
