+++ freedom-of-travel-impl/src/main/java/nc/project/service/ApartmentDetailsService.java
@@ -1,11 +1,11 @@
 package nc.project.service;
 
 import lombok.RequiredArgsConstructor;
-import nc.project.models.Apartment;
+import nc.project.jpa.entity.Apartment;
 import nc.project.models.ApartmentDetails;
-import nc.project.models.ImageApartment;
-import nc.project.repository.ApartmentRepository;
-import nc.project.repository.ImageApartmentRepository;
+import nc.project.jpa.entity.ImageApartment;
+import nc.project.jpa.repository.ApartmentRepository;
+import nc.project.jpa.repository.ImageApartmentRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
