+++ freedom-of-travel-impl/src/main/java/nc/project/service/HotelDetailsService.java
@@ -2,11 +2,11 @@ package nc.project.service;
 
 import lombok.RequiredArgsConstructor;
 
-import nc.project.models.Hotel;
+import nc.project.jpa.entity.Hotel;
 import nc.project.models.HotelDetails;
-import nc.project.models.ImageHotel;
-import nc.project.repository.HotelRepository;
-import nc.project.repository.ImageHotelRepository;
+import nc.project.jpa.entity.ImageHotel;
+import nc.project.jpa.repository.HotelRepository;
+import nc.project.jpa.repository.ImageHotelRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
