+++ freedom-of-travel-security/src/main/java/nc/project/services/UserDetailsImpl.java
@@ -3,8 +3,8 @@ package nc.project.services;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import nc.project.models.User;
-import nc.project.repository.UserRepository;
+import nc.project.jpa.entity.User;
+import nc.project.jpa.repository.UserRepository;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
